// apps/web/src/app/api/model/user.ts
/**
 * Talos API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface User {
    id: string;
    email: string;
    name: string;
    role: Role;
    /**
     * A deactivated user can no longer log in; already-issued JWTs remain valid until their 24h expiry.
     */
    active: boolean;
    createdAt: string;
    updatedAt: string;
}
export namespace User {
}